+++ viteclientts/src/components/LogoutButton.tsx
@@ -4,13 +4,18 @@ import { useNavigate } from 'react-router-dom';
 import { IconButton } from '@mui/material';
 import { ExitToApp } from '@mui/icons-material';
 import { useUserStore } from '../utils/store';
+import useSnackbar from '../utils/useSnackbar';
+import { useTranslation } from 'react-i18next';
 
 const LogoutButton: React.FC = () => {
   const navigate = useNavigate();
-  const { setUserLogin, clearJwtToken, setPreviousLastLogin, setHasVisitedBefore } = useUserStore();
+  const { setUserLogin, clearJwtToken, setPreviousLastLogin, setHasVisitedBefore, setLogoutInProgress } = useUserStore();
   const apiUrl = import.meta.env.VITE_API_URL;
+  const { showSnackbar } = useSnackbar();
+  const { t } = useTranslation();
 
   const handleLogout = async () => {
+    setLogoutInProgress(true);
     try {
       await axios.post(`${apiUrl}/logout`, {});
       setUserLogin(null);
@@ -18,6 +23,7 @@ const LogoutButton: React.FC = () => {
       setPreviousLastLogin(null);
       setHasVisitedBefore(null);
       navigate('/login');
+      showSnackbar(t('loggedOut'));
     } catch (error) {
       console.error('Logout error:', error);
     }